requirements.txt
# Camera RTSP Application Dependencies
# Python 3.8+ required

# Core GUI and Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0

# HTTP Client
requests>=2.31.0

# AI/ML Dependencies
insightface>=0.7.3
onnxruntime>=1.15.0
scikit-learn>=1.3.0

# Configuration and Environment
python-dotenv>=1.0.0

# Optional: GPU Support (uncomment if you have NVIDIA GPU)
# onnxruntime-gpu>=1.15.0

# Optional: Enhanced logging
# colorlog>=6.7.0

# Optional: Database support (if you want to use SQLite instead of JSON)
# sqlite3  # Built-in Python module

# Development Dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0

# System-specific dependencies
# Windows specific
# pywin32>=306; sys_platform == "win32"

# Linux specific
# python3-dev; sys_platform == "linux"

# macOS specific
# python3-dev; sys_platform == "darwin"